requirements.txt
Flask==3.0.0
pymysql==1.1.0
ultralytics==8.0.0
edge-tts==6.1.9
requests==2.31.0
opencv-python==4.8.1.78
Werkzeug==3.0.1
ollama==0.1.6
numpy==1.24.3
pyttsx3==2.90
geopy==2.4.0
Pillow==10.1.0
